Adding resource - create League

+++ src/main/java/leandroportfolio/league/resources/dto/CreateLeagueDTO.java
@@ -14,5 +14,5 @@ public class CreateLeagueDTO implements Serializable{
 	}
 	
 	public List<String> nicks;
-	public Long leagueTypeId;
+	public int leagueTypeId;
 }
